Diagnostics: Updated branch to trunk: - Fixed bug; Given parent IDs were not properly checked during import. - Updated the manual; added detail on installation requirements, and clarified that some extra fields during import are now optional. - Fixed bug; Importing a file with phenotype or screening data without a given individual ID, caused a fatal query error instead of an error message. - Display error message when filetype "Owned data download" is imported. - Fixed typo in the data importer. - Legend of a form field is available in data entry forms. Legend can now be viewed in via help icons. - Option added in the submission process to add a variant to a screening for when a submitter returned to the individual, or when the submission process is interrupted and continued through "Unfinished Submissions". - Fixed bug; When deleting all variants and associated data of a gene, deleting phenotypes while no individuals were found caused an error. - Added info table which explains the effect of deleting a gene. - Fixed bug; Solved bug with finding the proper path elements due to spaces in the LOVD installation path. - Fixed bug; When genes overlap, the protein change prediction predicted by the variant mapper was sometimes incorrect.
git-svn-id: https://humgenprojects.lumc.nl/svn/LOVD3/branches/NGS_diagnostics@509 a7477d18-a37e-46de-aef1-57953c184492

+++ src/genes.php
@@ -4,12 +4,13 @@
  * LEIDEN OPEN VARIATION DATABASE (LOVD)
  *
  * Created     : 2010-12-15
- * Modified    : 2014-08-06
- * For LOVD    : 3.0-11
+ * Modified    : 2014-12-11
+ * For LOVD    : 3.0-13
  *
  * Copyright   : 2004-2014 Leiden University Medical Center; http://www.LUMC.nl/
  * Programmers : Ing. Ivo F.A.C. Fokkema <I.F.A.C.Fokkema@LUMC.nl>
  *               Ing. Ivar C. Lugtenburg <I.C.Lugtenburg@LUMC.nl>
+ *               Msc. Daan Asscheman <D.Asscheman@LUMC.nl>
  *
  *
  * This file is part of LOVD.
@@ -427,6 +428,7 @@ if (PATH_COUNT == 1 && ACTION == 'create') {
                 lovd_writeLog('Event', LOG_EVENT, 'Created gene information entry ' . $_POST['id'] . ' (' . $_POST['name'] . ')');
 
                 // Make current user curator of this gene.
+                // DIAGNOSTICS: 
                 // $_DB->query('INSERT INTO ' . TABLE_CURATES . ' VALUES (?, ?, ?, ?)', array($_AUTH['id'], $_POST['id'], 1, 1));
 
                 // Add diseases.
@@ -828,18 +830,20 @@ if (PATH_COUNT == 2 && preg_match('/^[a-z][a-z0-9#@-]+$/i', rawurldecode($_PE[1]
                 $_BAR->setProgress(60);
                 $_BAR->setMessage('Deleting phenotypes...');
 
-                // Delete the Phenotypes! (NOTE: Again, just because I want the statistics...)
-                $q = $_DB->query('DELETE FROM ' . TABLE_PHENOTYPES . ' WHERE individualid IN (?' . str_repeat(', ?', count($aIndividuals) - 1) . ')', $aIndividuals);
-                $aDone['Phenotypes'] = $q->rowCount();
-                $nDone ++;
-                $_BAR->setProgress(80);
-                $_BAR->setMessage('Deleting individuals...');
+                if ($aIndividuals) {
+                    // Delete the Phenotypes! (NOTE: Again, just because I want the statistics...)
+                    $q = $_DB->query('DELETE FROM ' . TABLE_PHENOTYPES . ' WHERE individualid IN (?' . str_repeat(', ?', count($aIndividuals) - 1) . ')', $aIndividuals);
+                    $aDone['Phenotypes'] = $q->rowCount();
+                    $nDone ++;
+                    $_BAR->setProgress(80);
+                    $_BAR->setMessage('Deleting individuals...');
 
-                // And finally, delete the Individuals!
-                $q = $_DB->query('DELETE FROM ' . TABLE_INDIVIDUALS . ' WHERE id IN (?' . str_repeat(', ?', count($aIndividuals) - 1) . ')', $aIndividuals);
-                $aDone['Individuals'] = $q->rowCount();
-                $nDone ++;
-                unset($aIndividuals); // Save some memory.
+                    // And finally, delete the Individuals!
+                    $q = $_DB->query('DELETE FROM ' . TABLE_INDIVIDUALS . ' WHERE id IN (?' . str_repeat(', ?', count($aIndividuals) - 1) . ')', $aIndividuals);
+                    $aDone['Individuals'] = $q->rowCount();
+                    $nDone ++;
+                    unset($aIndividuals); // Save some memory.
+                }
             }
 
             $_BAR->setProgress(100);
@@ -959,6 +963,9 @@ if (PATH_COUNT == 2 && preg_match('/^[a-z][a-z0-9#@-]+$/i', rawurldecode($_PE[1]
 
     $_T->printHeader();
     $_T->printTitle();
+
+    lovd_showInfoTable('This will delete the ' . $zData['id'] . ' gene, all transcripts of this gene, and all annotations on variants specific for ' . $zData['id'] . '. The genomic variants and all individual-related information, including screenings, phenotypes and diseases, will not be deleted, so these might be left without a curator able to manage the data.<BR>
+                        <B>If you also wish to remove all information on individuals with variants in ' . $zData['id'] . ', first <A href="' . CURRENT_PATH . '?empty">empty</A> the gene database.</B>', 'warning');
 
     lovd_errorPrint();
